[FE] Refactor: RegisterForm - 중복된 폼 아이템 컴포넌트 분리
